requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-bootstrap3==10.0.1
django-extensions==2.1.1
django-filter==2.0.0
django-formtools==2.1
django-leaflet==0.24.0
#django-reversion==3.0.0
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
future==0.16.0
idna==2.7
jsonfield==2.0.2
oauthlib==2.1.0
pbkdf2==1.3
Pillow==5.2.0
py-ecc==1.4.3
pycparser==2.18
pycryptodome==3.6.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
repoze.lru==0.7
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
toolz==0.9.0
urllib3==1.23
django-braces==1.13.0
django-filters==0.2.1
weasyprint==0.42.3
django-markup-deprecated==0.0.3
pygments==2.2.0
docutils==0.12
django-chartit==0.2.9
openpyxl==2.5.9
datetime
django-hashedfilenamestorage==2.3
django-hijack
django-hijack-admin
#break
argcomplete==1.8.2
python-pptx==0.6.6
docx2txt==0.6
beautifulsoup4==4.6.0
xlrd==1.0.0
SpeechRecognition==3.7.1
django-reversion==3.0.3
pdfminer.six
odfpy==1.4.0
